Added Value Properties and Methods and Performance improvements
Added The 'Read Only' Property for the Value Classes.
Hard coded the Read Only Value for all Constant classes.
Added the To and From Variant Methods for value classes.
Implemented the Variant To and From Methods for all abstract types
Set all serialiseation methods to Reentrant
Removed redundant error cases from the serialisation methods and the Static Read and Static Write methods
Set the Static Read and write methods to be 'inline'